requirements.txt
mss
google-generativeai
requests
schedule
python-dotenv
rumps
Pillow
